+++ Frameworks/AVFoundation.framework/AVAssetResourceLoadingDataRequest.h
@@ -3,23 +3,23 @@
  */
 
 @interface AVAssetResourceLoadingDataRequest : NSObject {
-    AVAssetResourceLoadingDataRequestInternal *_dataRequest;
+    AVAssetResourceLoadingDataRequestInternal * _dataRequest;
 }
 
-@property (nonatomic, readonly) long long currentOffset;
+@property (nonatomic, readonly) int currentOffset;
 @property (nonatomic, readonly) int requestedLength;
-@property (nonatomic, readonly) long long requestedOffset;
+@property (nonatomic, readonly) int requestedOffset;
 @property (nonatomic, readonly) BOOL requestsAllDataToEndOfResource;
 
 - (id)_loadingRequest;
-- (long long)currentOffset;
+- (int)currentOffset;
 - (void)dealloc;
 - (id)description;
 - (void)finalize;
 - (id)init;
-- (id)initWithLoadingRequest:(id)arg1 requestedOffset:(long long)arg2 requestedLength:(long long)arg3 requestsAllDataToEndOfResource:(BOOL)arg4 canSupplyIncrementalDataImmediately:(BOOL)arg5;
+- (id)initWithLoadingRequest:(id)arg1 requestedOffset:(int)arg2 requestedLength:(int)arg3 requestsAllDataToEndOfResource:(BOOL)arg4 canSupplyIncrementalDataImmediately:(BOOL)arg5;
 - (int)requestedLength;
-- (long long)requestedOffset;
+- (int)requestedOffset;
 - (BOOL)requestsAllDataToEndOfResource;
 - (void)respondWithData:(id)arg1;
