+++ Frameworks/AVFoundation.framework/AVAssetWriterInputNoMorePassesHelper.h
@@ -3,10 +3,10 @@
  */
 
 @interface AVAssetWriterInputNoMorePassesHelper : AVAssetWriterInputHelper {
-    AVAssetWriterInputWritingHelper *_writingHelper;
+    AVAssetWriterInputWritingHelper * _writingHelper;
 }
 
-- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg2;
+- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { int x1; int x2; unsigned int x3; int x4; })arg2;
 - (int)appendSampleBuffer:(struct opaqueCMSampleBuffer { }*)arg1 error:(id*)arg2;
 - (BOOL)canPerformMultiplePasses;
 - (id)currentPassDescription;
