+++ Frameworks/AVFoundation.framework/AVAssetWriterInputPixelBufferAdaptor.h
@@ -3,7 +3,7 @@
  */
 
 @interface AVAssetWriterInputPixelBufferAdaptor : NSObject <AVKeyPathDependencyHost, AVWeakObservable> {
-    AVAssetWriterInputPixelBufferAdaptorInternal *_internal;
+    AVAssetWriterInputPixelBufferAdaptorInternal * _internal;
 }
 
 @property (nonatomic, readonly) AVAssetWriterInput *assetWriterInput;
@@ -17,7 +17,7 @@
 + (id)assetWriterInputPixelBufferAdaptorWithAssetWriterInput:(id)arg1 sourcePixelBufferAttributes:(id)arg2;
 
 - (void)addCallbackToCancelDuringDeallocation:(id)arg1;
-- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg2;
+- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { int x1; int x2; unsigned int x3; int x4; })arg2;
 - (id)assetWriterInput;
 - (void)dealloc;
 - (void)declareKeyPathDependenciesWithRegistry:(id)arg1;
