+++ Frameworks/AVFoundation.framework/AVAssetWriterWritingHelper.h
@@ -3,10 +3,10 @@
  */
 
 @interface AVAssetWriterWritingHelper : AVAssetWriterHelper <AVAssetWriterFigAssetWriterNotificationHandlerDelegate> {
-    struct OpaqueFigAssetWriter { } *_figAssetWriter;
-    NSObject<OS_dispatch_queue> *_figAssetWriterAccessQueue;
-    AVAssetWriterFigAssetWriterNotificationHandler *_notificationHandler;
-    BOOL _startSessionCalled;
+    struct OpaqueFigAssetWriter { } * _figAssetWriter;
+    NSObject<OS_dispatch_queue> * _figAssetWriterAccessQueue;
+    AVAssetWriterFigAssetWriterNotificationHandler * _notificationHandler;
+    BOOL  _startSessionCalled;
 }
 
 @property (readonly, copy) NSString *debugDescription;
@@ -23,14 +23,14 @@
 - (void)cancelWriting;
 - (void)dealloc;
 - (void)didReceiveFigAssetWriterNotificationWithSuccess:(BOOL)arg1 error:(id)arg2;
-- (void)endSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
+- (void)endSessionAtSourceTime:(struct { int x1; int x2; unsigned int x3; int x4; })arg1;
 - (id)figTrackReferences;
 - (void)finalize;
 - (void)finishWriting;
 - (void)finishWritingWithCompletionHandler:(id /* block */)arg1;
 - (id)initWithConfigurationState:(id)arg1;
 - (id)initWithConfigurationState:(id)arg1 error:(id*)arg2;
-- (void)startSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
+- (void)startSessionAtSourceTime:(struct { int x1; int x2; unsigned int x3; int x4; })arg1;
 - (int)status;
 
 @end
