+++ Frameworks/AVFoundation.framework/AVAsynchronousVideoCompositionRequest.h
@@ -3,24 +3,24 @@
  */
 
 @interface AVAsynchronousVideoCompositionRequest : NSObject <NSCopying> {
-    AVAsynchronousVideoCompositionRequestInternal *_internal;
+    AVAsynchronousVideoCompositionRequestInternal * _internal;
 }
 
-@property (nonatomic, readonly) struct { long long x1; int x2; unsigned int x3; long long x4; } compositionTime;
+@property (nonatomic, readonly) struct { int x1; int x2; unsigned int x3; int x4; } compositionTime;
 @property (nonatomic, readonly) AVVideoCompositionRenderContext *renderContext;
 @property (nonatomic, readonly) NSArray *sourceTrackIDs;
 @property (nonatomic, readonly) <AVVideoCompositionInstruction> *videoCompositionInstruction;
 
 - (void)_willDeallocOrFinalize;
 - (struct OpaqueFigVideoCompositorFrame { }*)compositionFrame;
-- (struct { long long x1; int x2; unsigned int x3; long long x4; })compositionTime;
+- (struct { int x1; int x2; unsigned int x3; int x4; })compositionTime;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (void)dealloc;
 - (void)finalize;
 - (void)finishCancelledRequest;
 - (void)finishWithComposedVideoFrame:(struct __CVBuffer { }*)arg1;
 - (void)finishWithError:(id)arg1;
-- (id)initUsingSession:(id)arg1 withRenderContext:(id)arg2 compositionFrame:(struct OpaqueFigVideoCompositorFrame { }*)arg3 atTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg4 usingSources:(id)arg5 instruction:(id)arg6;
+- (id)initUsingSession:(id)arg1 withRenderContext:(id)arg2 compositionFrame:(struct OpaqueFigVideoCompositorFrame { }*)arg3 atTime:(struct { int x1; int x2; unsigned int x3; int x4; })arg4 usingSources:(id)arg5 instruction:(id)arg6;
 - (id)renderContext;
 - (struct __CVBuffer { }*)sourceFrameByTrackID:(int)arg1;
 - (id)sourceTrackIDs;
