+++ Frameworks/AVFoundation.framework/AVAudioMixSweepFilterEffectParameters.h
@@ -3,13 +3,13 @@
  */
 
 @interface AVAudioMixSweepFilterEffectParameters : AVAudioMixEffectParameters {
-    AVAudioMixSweepFilterEffectParametersInternal *_sweepFilterAudioEffect;
+    AVAudioMixSweepFilterEffectParametersInternal * _sweepFilterAudioEffect;
 }
 
-@property (nonatomic, readonly) float bypassThreshold;
+@property (nonatomic, readonly) double bypassThreshold;
 @property (nonatomic, readonly) int maximumCutOffFrequency;
 @property (nonatomic, readonly) int minimumCutOffFrequency;
-@property (nonatomic, readonly) float sweepValue;
+@property (nonatomic, readonly) double sweepValue;
 
 + (id)sweepFilterEffectParametersWithMinimumCutOffFrequency:(unsigned int)arg1 maximumCutOffFrequency:(unsigned int)arg2 bypassThreshold:(float)arg3 sweepValue:(float)arg4;
